Move all the logic again to a new ViewController :(
